redaccess_service/api/serializers/company_serializer.py
from rest_framework import serializers

from api.models.company import Company

class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company
        fields = '__all__'