redaccess_service/requirements.txt
django
djangorestframework
djangorestframework-dataclasses
dataclasses-json
django-rest-authtoken
django-cors-headers